+++ tests/sdk/nodejs/examples/helm-release-redis/step1/index.ts
@@ -1,0 +1,47 @@
+import * as pulumi from "@pulumi/pulumi";
+import * as k8s from "@pulumi/kubernetes";
+import * as random from "@pulumi/random";
+
+const redisPassword = new random.RandomPassword("password", {length: 15});
+
+const nsName = new random.RandomPet("test");
+const namespace = new k8s.core.v1.Namespace("release-ns", {
+    metadata: {
+        name: nsName.id
+    }
+});
+
+// Validates fix for https://github.com/pulumi/pulumi-kubernetes/issues/1933
+function values(password: pulumi.Output<string>): pulumi.Input<{ [key: string]: any }> {
+    return pulumi.output({
+        cluster: {
+            enabled: true,
+            slaveCount: 1,
+        },
+        global: {
+            redis: {
+                password: password,
+            }
+        },
+        rbac: {
+            create: true,
+        },
+        service: {
+            type: "ClusterIP"
+        }
+    });
+};
+
+const release = new k8s.helm.v3.Release("redis", {
+    chart: "redis",
+    repositoryOpts: {
+        repo: "https://charts.bitnami.com/bitnami",
+    },
+    namespace: namespace.metadata.name,
+    values: values(redisPassword.result),
+});
+
+
+const srv = k8s.core.v1.Service.get("redis-master-svc", pulumi.interpolate`${release.status.namespace}/${release.status.name}-master`);
+export const redisMasterClusterIP = srv.spec.clusterIP;
+export const status = release.status.status;